chromium/third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-flexbox/dynamic-grid-flex-abspos-expected.txt

This is a testharness.js-based test.
PASS .relpos 1
PASS .relpos 2
PASS .relpos 3
FAIL .relpos 4 assert_equals: 
<div class="relpos flexbox" data-expected-height="20">
    <div class="abspos">Test</div>
    <div style="height: 20px;"></div>
    <div style="height: 50%" data-expected-height="10"></div>
  </div>
height expected 10 but got 0
PASS .relpos 5
PASS .relpos 6
PASS .relpos 7
FAIL .relpos 8 assert_equals: 
<div class="relpos flexbox" data-expected-height="20">
    <div class="abspos" style="border-width: 10px;">Test</div>
    <div style="height: 20px;"></div>
    <div style="height: 50%" data-expected-height="10"></div>
  </div>
height expected 10 but got 0
Harness: the test ran to completion.
